Fixup build scripts a bit
Remove bash dependency and just add some error checks, nothing fancy

+++ release_build.sh
@@ -1,8 +1,21 @@
+#!/bin/sh
+
 # this is used to create an rsyslog release tarball.
 # it must be executed from the root of the rsyslog-doc
 # project.
-rm -rf ./build
-sphinx-build -b html source build
-rm rsyslog-doc.tar.gz
-tar -czf rsyslog-doc.tar.gz build source LICENSE README.md build.sh
-cd ..
+
+docfile=rsyslog-doc.tar.gz
+
+[ ! -d ./build ] || rm -rf ./build
+
+sphinx-build -b html source build || {
+	echo "sphinx-build failed... aborting"
+	exit 1
+}
+
+[ ! -e ./$docfile ] || rm -rf ./$docfile
+
+tar -czf $docfile build source LICENSE README.md build.sh || {
+	echo "Failed to create rsyslog-doc.tar.gz tarball..."
+	exit 1
+}
